# More efficient storage storage format