--- a/dart/.gitignore
+++ b/dart/.gitignore
@@ -1,0 +1,1 @@
+.dart_tool