This changes `ps_oplog` to only store PUT operations, roughly in the format described [here](https://github.com/powersync-ja/powersync-service/blob/main/docs/bucket-properties.md#reducing-buckets).

Specific changes on ps_oplog:
1. Remove `superseded` column.
2. Remove `op` column.
3. Change `bucket` column from TEXT (bucket name) to INTEGER (bucket id).

Instead of storing pending REMOVE operations in `ps_oplog`, it is now stored in a dedicated `ps_updated_rows` table. Only the row type and id are relevant here - it indicates rows which may need to be re-synced from the local bucket data.

This has some advantages:
1. Reduced storage requirements.
2. No need for a slow "compact" operation to delete old REMOVE operations.
3. Simplifies deleting of buckets.

The operations `delete_pending_buckets` and `clear_remove_ops` are now no-ops, and may be removed from the respective SDKs.

## Benchmarks

Flutter desktop on linux, 500k PUT operations in a single `global[]` bucket, time initial sync and database size.

Before: 56s total, of which 18s is spent in `sync_local`. 401MB database.
After: 57s total, of which 19s is spent in `sync_local`. 390MB database.

Conclusion: Does not significantly affect initial sync time. Does slightly reduces database size. May have a bigger effect when syncing many REMOVE operations, when deleting buckets, or when there are long bucket names.

## Tests

This now adds dart-based tests to test the extension - specifically the schema migrations. This could be any high-level language - just used Dart because the tests were already written in Dart.

Other tests may be moved from the powersync.dart repo to here in the future.